fix[all]: thêm xác thực lần 1

+++ backend/src/main/java/com/react_online_judge/backend/service/ProblemService.java
@@ -1,10 +1,12 @@
 package com.react_online_judge.backend.service;
 
+import com.nimbusds.jose.JOSEException;
 import com.react_online_judge.backend.dto.request.ProblemCreationRequest;
 import com.react_online_judge.backend.dto.request.ProblemUpdateRequest;
 import com.react_online_judge.backend.dto.response.ProblemResponse;
 import com.react_online_judge.backend.entity.Announcement;
 import com.react_online_judge.backend.entity.Problem;
+import com.react_online_judge.backend.entity.User;
 import com.react_online_judge.backend.exception.AppException;
 import com.react_online_judge.backend.exception.ErrorCode;
 import com.react_online_judge.backend.mapper.ProblemMapper;
@@ -17,6 +19,7 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
 
+import java.text.ParseException;
 import java.time.LocalDateTime;
 import java.util.List;
 
@@ -28,6 +31,8 @@ public class ProblemService {
     ProblemRepository problemRepository;
     ProblemMapper problemMapper;
     AnnouncementRepository announcementRepository;
+    private final AuthenticateService authenticateService;
+
     public ProblemResponse getProblemById(Long id) {
         Problem problem = problemRepository.findById(id).orElseThrow(() -> new AppException(ErrorCode.PROBLEM_NOT_EXISTED));
         return problemMapper.toProblemResponse(problem);
@@ -44,12 +49,28 @@ public class ProblemService {
         List<Problem> problems = problemRepository.findByContestId(contestId);
         return problemMapper.toProblemResponseList(problems);
     }
+    public List<ProblemResponse> getProblemsByCreator(String token) {
+        try {
+            User user = authenticateService.getUserFromToken(token);
+            List<Problem> problems = problemRepository.findAllByAuthor(user.getUsername());
+            return problemMapper.toProblemResponseList(problems);
+        } catch (ParseException | JOSEException e) {
+            throw new AppException(ErrorCode.UNAUTHENTICATED);
+        }
+    }
     public List<ProblemResponse> getAllPublicProblems() {
         List<Problem> problems = problemRepository.findAllByPublicFlag(true);
         return problemMapper.toProblemResponseList(problems);
     }
-    public ProblemResponse createProblem(ProblemCreationRequest request) {
+    public ProblemResponse createProblem(String token, ProblemCreationRequest request) {
         Problem problem = problemMapper.toProblem(request);
+        User user;
+        try {
+            user = authenticateService.getUserFromToken(token);
+            problem.setAuthor(user.getUsername());
+        } catch (ParseException | JOSEException e) {
+            throw new AppException(ErrorCode.UNAUTHENTICATED);
+        }
         try {
             problem = problemRepository.save(problem);
             if (problem.isPublicFlag()) {
@@ -66,9 +87,18 @@ public class ProblemService {
             throw new AppException(ErrorCode.PROBLEM_EXISTED);
         }
     }
-    public ProblemResponse updateProblem(Long id, ProblemUpdateRequest request) {
+    public ProblemResponse updateProblem(String token, Long id, ProblemUpdateRequest request) {
         Problem problem = problemRepository.findById(id).orElseThrow(() -> new AppException(ErrorCode.PROBLEM_NOT_EXISTED));
         problemMapper.updateProblem(problem, request);
+        User user;
+        try {
+            user = authenticateService.getUserFromToken(token);
+            if (!problem.getAuthor().equals(user.getUsername())) {
+                throw new AppException(ErrorCode.UNAUTHENTICATED);
+            }
+        } catch (ParseException | JOSEException e) {
+            throw new RuntimeException(e);
+        }
         try {
             problemRepository.save(problem);
             return problemMapper.toProblemResponse(problem);
@@ -76,7 +106,17 @@ public class ProblemService {
             throw new AppException(ErrorCode.PROBLEM_EXISTED);
         }
     }
-    public void deleteProblem(Long id) {
+    public void deleteProblem(String token, Long id) {
+        User user;
+        try {
+            user = authenticateService.getUserFromToken(token);
+        } catch (ParseException | JOSEException e) {
+            throw new AppException(ErrorCode.UNAUTHENTICATED);
+        }
+        Problem problem = problemRepository.findById(id).orElseThrow(() -> new AppException(ErrorCode.PROBLEM_NOT_EXISTED));
+        if (!problem.getAuthor().equals(user.getUsername())) {
+            throw new AppException(ErrorCode.UNAUTHENTICATED);
+        }
         problemRepository.deleteById(id);
     }
 }
